Introduce asynchronous datafeed listening service
The current datafeed listening service uses requests to wait on
datafeed/{id}/read until a response is received, handle that and then
wait for further events. This means DatafeedEventService.read_datafeed
is a blocking method and nothing outside of handling events can be
performed by the program. It also means that messages must be received
and handled in order, and no further messages can be processed until
the current one is done. This limitation could be handled by threads
but they are error-prone in Python, and it's not clear if the codebase
is threadsafe.

Instead this commit adds another class, AsyncDataFeedEventService that
uses aiohttp and asyncio to listen to the datafeed in an asynchronous
manner, meaning other operations can be queued on the event loop. Since
the event loop runs in on the main thread, and typically bots are
IO-bound, this allows for a model where bots can receive events,
dispatch them to other async methods and quickly return to responding
to messages. The original datafeed reading service can still be used,
but for the price of some added complexity the asychronous version
allows for far higher throughput and responsiveness.

Several other additions have been made:
  - Update the error handling to hand 400 and >=500 errors from
    the datafeed, this now matches the Java implementation
  - Add a timeout to datafeed errors
  - Add the healthcheck client
  - Improve configuration loading to be less dependent on current
    directory and clearer about entries that don't exist
  - Add a testing framework for the datafeed
  - Add missing dependencies to setup.py