PLAT-10225: Fix datafeed loop crash for events without id
A datafeed event does not always have an id (for instance when external
users join a room) therefore we must handle that. The access to the id
is replaced with .get() to avoid the runtime error and if the id is
missing a random one is set for tracing purpose.

Added a unit test for that using IsolatedAsyncioTestCase, it is a bit
complex due to the async nature of the code but it does run part of the
datafeed loop.