PLAT-10245: Support CircleCI build for master branch
In the 2.0 branch we introduced CircleCI and PR builders, just to make
it work with the master branch, here we add a simple build file just
running the tests (the bare minimum, the 2.0 branch supports Snyk, test
coverage and reports).

3.6 cannot be used because we use unittest.async_case.
Caching with pyenv was not working (dependencies problem) so it is not
implemented (we do not expect a lot of executions).