Message Service implementation.
Create MultiAttachmentsMessageApi class inherited MessagesApi allow bot to call the api which will send a message with multiple attachments and files as a list of files.

Create MessageService class wrapping all the service api related to message service which are documented in the https://developers.symphony.com/restapi/reference#messages-v4

Create method in the SymphonyBdk class to initialize the MessageService instance and provide it to bot developer to work with message service.

Create an example which shows how a bot interact with the message service.

Fix wrong path for agent api client.