User Service implementation.
Create UserService class wrapping all the service api related to message service which are documented in the developers.symphony.com site.

Create method in the SymphonyBdk class to initialize the UserService instance and provide it to bot developer to work with message service.

Create an example which shows how a bot interact with the user service.

Unittest added.

Documentation added.