PLAT-10459: Logging configuration example
As part of the datafeed example a sample logging configuration is
provided, sending logs to the console and to rotating files. This is
mostly here to illustrate how logging can be used in a bot.

Also handled KeyboardInterruptException in the example to avoid the
stacktrace when we quit the application.

The example has been rewritten as a script, if you are ok with that I'll
move other examples as well. It is shorter and easier to understand.

poetry build was failing so the package has been configured properly and
the package's name changed to what is actually published on pypi (might
change later).