PLAT-10343: Datafeed V2 service implementation (#124)
* Datafeed Loop v2 implementation

In the BdkDatafeedConfig class, check if the `datafeed.version` configuration is
specified. If not, the datafeed version equals `v1`

In the ServiceFactory class, check if the datafeed version is configured as `v2`, return
the DatafeedLoopV2 instance. Otherwise, the DatafeedLoopV1 will be used by default.

Implementation the datafeed v2 loop service in DatafeedLoopV2 class

* Unittest added

* Update markdown documentation