PLAT-10529 Ability to set private key and certificate content programmatically
Private key and Certificate were able to only be loaded from the config file when initializing bdk. In this PR, we update bot config to be able to set private key and certificate after bdk being initialized.
BdkAuthenticationConfig constructor overrides both private key and certificate even when only one is provided, the missing one is overriden to the default value and thus its content is lost. This is why two setters have been added to the class to be able to set one independantly of the other.