PLAT-10643: Load certs from system store
In #122 we enable loading a custom cert store for the HTTP client. This
broked using pods with valid certs because the system certs are no
longer loaded. This change loads them even all the time (and we might
add custom ones on top of that).

Also sync the rest.mustache template used for code generation.