PLAT-10895-Implement built-in help command
Goal of this PR is to provide the implementation of a build-in help command with lists all the activity registered in the ActivityRegistry.
For each command found the following format will be used " - command_name - description (mention required/mention not required)" (example "/hello - command to say hello (mention required)")
This help command can also be overriden by manually implementing a new SlashCommandActivity, the code will make sure that any duplicates will be removed.