Retry on client/timeout errors
It looks like the existing logic to retry on timeouts or client errors
was not working as expected. For instance the DF loop would exit on
timeouts.

The aiohttp client can raise different errors
(https://docs.aiohttp.org/en/stable/client_reference.html#client-exceptions)
and also asyncio.TimeoutError.

I tried locally by misconfiguring the hostname (it cannot be resolved)
and the port (triggers a timeout).

However this means that retries will be performed upon startup even if
the hostname cannot be resolved for instance. Which is not 100% like the
Java BDK who has a different logic for authentication and datafeed
retries. The problem is how sessions are refreshed, here lazily so we
have two nested @retry calls making it difficult to have different
strategies.

Fixes #256