Starting Datafeed v1 deprecation process (#270)
* Starting Datafeed v1 deprecation process
The datafeed 1 service will be fully replaced by the datafeed 2 service in the future. Please consider migrating over to datafeed 2 APIs /agent/v5/datafeeds. For more information on the timeline as well as on the benefits of datafeed 2, please reach out to your Technical Account Manager or to our developer documentation https://docs.developers.symphony.com/building-bots-on-symphony/datafeed

This PR includes:
- if datafeed version is not defined in the config, the BDK will use DFv2 by default
- added deprecation notice when DFv1 is used