Implementation of Datahose loop.
Implemented Datahose loop. It is based on an ack_id like Datafeed V2, that is why a refactor is done in order to get the common logic in a new abstract class abstract_ackId_event_loop, that in turn extends abstract_datafeed_loop and defines common intereface.
Datahose loop is not supposed to be used to react to activities but it is meant to subscribe to different real time events for monitoring and scanning purposes, that is why the interface does not allow to register activities.

The generated api/code has been updated to support last changes.