Implementation of Datahose loop (#281)
Implementation of Datahose loop.

Implemented Datahose loop. It is based on an ack_id like Datafeed V2, that is why a refactor is done in order to get the common logic in a new abstract class abstract_ackId_event_loop, that in turn extends abstract_datafeed_loop and defines common interface.
Datahose loop is not supposed to be used to react to activities but it is meant to subscribe to different real time events for monitoring and scanning purposes, that is why the interface does not allow to register activities.

Introduced abstract_datahose_loop interface to distinguish datahose from datafeed implementation from a developer point of view, given both are instances of abtract_datafeed_loop

I also made recreate_datafeed and prepare_datafeed not abstract in abstract_datafeed_loop.
Datahose extends abstract_datafeed_loop but it should not implement those methods as they are specific to DF1 and DF2. They are not abstarct methods anymore and they are declared and implemented in the concrete classes for DFV1 and DFV2 loops.