#511 Add accessible slash command arguments (#284)
* #511 Add accessible slash command arguments

In this commit, we made slash command support string/cashtag/hashtag/mention arguments. We introduced slashCommandPattern class to represent the pattern to match user input.
When user input matches the pattern provided in the slash command, we add arguments values in the commandContext in order to be used in the command callback logic.
When a command required bot mention, we check bot user id instead of the display name.
The command pattern is tokenised and user input as well. The matching logic consists of comparing tokens values.

Bot user id loaded in a lazy way during runtime, using a lambda function.

Closes #238