#295 Bug fix: presence not working with DF2 When using DF2, the bot presence is handled in Agent side. The datafeed ackId has been previously updated to include a boolean field update_presence in addition to the ackId string. This boolean field allows Agent to update bot presence along with the DF heartbeat. The field default value is True and cannot be changed by the bot (it is only customisable via API calls).
The issue preventing the presence update was happening because the default value of update_presence, which is True, was generated as None in BDK Python, which in turn is mapped to False in Agent side. Consequently, Agent was skipping bot presence update.

In this commit, we updated the generated code to use the latest that we changed to set the default value for optiona vars, if provided, otherwise None is used. Previously, None was always set in any case.