Fix Datahose loop (#311)
In the new update of Datahose api, the field filters has been renamed to eventTypes. In this commit, we apply this update in BDK Python.
Datahoseloop should not retry on client exceptions, only minor ones 500, 401 and 429 and timeouts.
The generated code has been updated as well according to the last changes in symphony-api-specs.