Not use bot username when Datafeed tag is not set (#320)
The Datafeed tag is an internal feature and the tag property was optional in the configuration. So far, when no tag is configured, the bot username was used as a tag. This leads to production issues when the tag is an old hanging one. In this PR, we get ride of the tag in Datafeed calls and configuration.
For the same occasion, we corrected the Datafeed documentation with some links returning a 404 page error.